+++ openmp/src/game_of_life.c
@@ -1,7 +1,7 @@
 #include "game_of_life.h"
 
 
-void init_gol(gol_t* gol, const uint32_t grid_size, const float density) {
+void init_gol(gol_t* gol, const uint64_t grid_size, const float density) {
     // size of the grid + 2 for the ghost cells
     gol->size = grid_size + 2;
     size_t size = (gol->size) * (gol->size) * sizeof(bool);
@@ -13,14 +13,14 @@ void init_gol(gol_t* gol, const uint32_t grid_size, const float density) {
 }
 
 void init_grid(const gol_t* gol, const float density) {
-    for (uint32_t i = 1; i < gol->size - 1; i++) {
-        for (uint32_t j = 1; j < gol->size - 1; j++) {
+    for (uint64_t i = 1; i < gol->size - 1; i++) {
+        for (uint64_t j = 1; j < gol->size - 1; j++) {
             gol->current[idx(gol, i, j)] = (((float) rand()) / RAND_MAX) < density;
         }
     }
 }
 
-uint8_t count_alive_neighbors(const gol_t* gol, const uint32_t i, const uint32_t j) {
+uint8_t count_alive_neighbors(const gol_t* gol, const uint64_t i, const uint64_t j) {
     return  gol->current[idx(gol, i - 1, j - 1)] + gol->current[idx(gol, i - 1, j)] + gol->current[idx(gol, i - 1, j + 1)] +
             gol->current[idx(gol, i, j - 1)    ]                                    + gol->current[idx(gol, i, j + 1)    ] +
             gol->current[idx(gol, i + 1, j - 1)] + gol->current[idx(gol, i + 1, j)] + gol->current[idx(gol, i + 1, j + 1)];
@@ -34,8 +34,8 @@ void swap_grids(gol_t* gol) {
 
 void step(const gol_t* gol) {
 #pragma omp parallel for collapse(2)
-    for (uint32_t i = 1; i < gol->size - 1; i++) {
-        for (uint32_t j = 1; j < gol->size - 1; j++) {
+    for (uint64_t i = 1; i < gol->size - 1; i++) {
+        for (uint64_t j = 1; j < gol->size - 1; j++) {
 
             uint8_t alive_neighbors = count_alive_neighbors(gol, i, j);
 
@@ -53,14 +53,14 @@ void step(const gol_t* gol) {
 void fill_ghost_cells(const gol_t* gol) {
     // Left and right borders
 #pragma omp parallel for
-    for (uint32_t i = 1; i < gol->size - 1; i++) {
+    for (uint64_t i = 1; i < gol->size - 1; i++) {
         gol->current[idx(gol, i, 0)] = gol->current[idx(gol, i, gol->size - 2)];
         gol->current[idx(gol, i, gol->size - 1)] = gol->current[idx(gol, i, 1)];
     }
 
     // Top and bottom borders
 #pragma omp parallel for
-    for (uint32_t j = 0; j < gol->size; j++) {
+    for (uint64_t j = 0; j < gol->size; j++) {
         gol->current[idx(gol, 0, j)] = gol->current[idx(gol, gol->size - 2, j)];
         gol->current[idx(gol, gol->size - 1, j)] = gol->current[idx(gol, 1, j)];
     }
